+++ var_stats_qc/qc.py
@@ -216,6 +216,10 @@ def compute_relatedness_many(connection, variant_model, cases, min_depth=7, n_si
     ibs1 = {p: 0 for p in sample_pairs}
     ibs2 = {p: 0 for p in sample_pairs}
 
+    # Return defaults if project is empty
+    if result.rowcount == 0:
+        return het, het_shared, ibs0, ibs1, ibs2
+
     # Iterate over genotypes of pseudo-autosomal regions on chrX.
     kept = 0
     for row in result.fetchall():
@@ -228,6 +232,8 @@ def compute_relatedness_many(connection, variant_model, cases, min_depth=7, n_si
             sample: _normalize_gt(genotype.get(sample, {}).get("gt", "./.")) for sample in samples
         }
         nocalls = [gt for gt in gts if gt == "./."]
+        if not gts:
+            continue
         if len(nocalls) / len(gts) > 0.1 and len(gts) - len(nocalls) < 5:
             continue  # skip, too few calls
         # Skip if depth not sufficient.